Use lazy text to read files
The migration to strict text made `vgrep` hang whenever a large file was
loaded to the pager. Strict text is suitable for representing individual
lines, but not for representing a huge file. The solution is to read the
file using lazy text, split thc content into lines, and then convert the
lines to strict text.

Fixes #30.