Fix: When invoked without arguments in the middle of a pipe, just pipe stdin to stdout
This happens for example when the Git alias `git vgrep` is set:

    git config --global alias.vgrep '!__git_vgrep () { git grep --color=always "$@" | vgrep; }; __git_vgrep'

Then a problem arises when piping the output of `git vgrep` to another command,

    git vgrep foo        # Works
    git vgrep foo | cat  # Fails

because vgrep tries to invoke `grep` without arguments. The solution is to allow
invoking vgrep in the middle of a pipe without arguments, and just piping the
stdin to stdout.